renamed events.js and updated import to main.js
